test update to skip benchmarking, car and carhandler updates to have cars wait in a line behind a stopped car in sequence.

+++ src/classes/CarHandler.cpp
@@ -189,7 +189,7 @@ void CarHandler::handleStop(int index) {
             Car* wait = me->getWait();
             pair<float, float> myWay = me->getWaypoint();
             pair<float, float> theirWay = closest->getWaypoint();
-            int diff = (myWay - theirWay).first + (myWay - theirWay).second;
+            int diff = (int)((myWay - theirWay).first + (myWay - theirWay).second);
             if (abs(diff) <= 2) {
                 closest->waitBehind(me);
                 me->setBehind(true);
